Liveness and readiness logic change
