+++ src/main/java/pl/lodz/p/it/ssbd2021/ssbd06/healthcheck/LivenessHealthCheck.java
@@ -5,7 +5,6 @@ import org.eclipse.microprofile.health.HealthCheckResponse;
 import org.eclipse.microprofile.health.Liveness;
 
 import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
 import java.lang.management.ManagementFactory;
 import java.lang.management.MemoryMXBean;
 
@@ -13,30 +12,20 @@ import java.lang.management.MemoryMXBean;
 @Liveness
 public class LivenessHealthCheck implements HealthCheck {
 
-//    private final MemoryMXBean memoryMXBean = ManagementFactory.getMemoryMXBean();
-
-    @Inject
-    private ReadinessKeeper keeper;
-
-//    @Override
-//    public HealthCheckResponse call() {
-//        return HealthCheckResponse.named("heap-memory")
-//                .status(getUsedMemory() < 0.9 * getMaxMemory())
-//                .build();
-//    }
+    private final MemoryMXBean memoryMXBean = ManagementFactory.getMemoryMXBean();
 
     @Override
     public HealthCheckResponse call() {
-        return HealthCheckResponse.named("is-app-up")
-                .status(keeper.isClicked())
+        return HealthCheckResponse.named("heap-memory")
+                .status(getUsedMemory() < 0.9 * getMaxMemory())
                 .build();
     }
 
-//    private long getUsedMemory() {
-//        return memoryMXBean.getHeapMemoryUsage().getUsed();
-//    }
-//
-//    private long getMaxMemory() {
-//        return memoryMXBean.getHeapMemoryUsage().getMax();
-//    }
+    private long getUsedMemory() {
+        return memoryMXBean.getHeapMemoryUsage().getUsed();
+    }
+
+    private long getMaxMemory() {
+        return memoryMXBean.getHeapMemoryUsage().getMax();
+    }
 }
